+++ client/src/test/java/Client/data/ConvertCsvToJsonTest.java
@@ -1,6 +1,6 @@
 package Client.data;
 
-import de.qaware.chronix.client.benchmark.util.CsvImporter;
+import de.qaware.chronix.shared.QueryUtil.CsvImporter;
 
 import java.io.File;
 import java.util.ArrayList;
