+++ client/src/test/java/Client/data/ReadMetaDataFromJsonTest.java
@@ -1,6 +1,6 @@
 package Client.data;
 
-import de.qaware.chronix.client.benchmark.util.JsonTimeSeriesHandler;
+import de.qaware.chronix.shared.QueryUtil.JsonTimeSeriesHandler;
 import de.qaware.chronix.database.TimeSeriesMetaData;
 
 import java.util.ArrayList;
